backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
joblib==1.3.2
numpy==1.24.3
razorpay==1.4.2
apscheduler==3.10.4